Easier API for working with the GPIO breakout header on the interface board
The GPIO module defines GPIOPin and HeaderPin classes.  The former identifies
pins by index in the GPIO header in the Q2W breakout board (0..7).  The
latter identifies pins by the index in the Raspberry Pi on-board header.

For backward compatability, Pin is a synonym for HeaderPin.

There is no PinBank abstraction yet.